+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.0.23] - 2023-01-15
+
+### Added
+
+- Various values related to shopkeeper aggro in Debug->Global State menu
+- Held Item Metadata to Player tab
+- Hired Hand Count to Player tab
+
 ## [0.0.22] - 2023-01-10
 
 ### Added